requirements.txt
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
python-dotenv==1.2.1
psycopg2-binary==2.9.11
Werkzeug==3.1.3
gunicorn==22.0.0
